==> Voter_Classifications/3_feature_comb.ipynb <==
# Feature Combination

import pandas as pd
import numpy as np
import xgboost as xgb
import seaborn as sns
from scipy.stats import skew, boxcox
from scipy import sparse
from sklearn.model_selection import KFold,GridSearchCV
from sklearn.ensemble import ExtraTreesRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn import metrics, preprocessing
import time
import matplotlib.pyplot as plt
%matplotlib inline

## Load data

train_label = pd.read_csv('../ProjectFiles/File1.csv')
train_label_size = train_label.shape[0]
train = pd.read_csv('../ProjectFiles/File2.csv')
train_size = train.shape[0]
test = pd.read_csv('../ProjectFiles/File3.csv')
test_size = test.shape[0]

df_test = pd.concat([test,train], axis=1)
df_test = df_test.loc[:,~df_test.columns.duplicated()]
df_test.shape

df_full = pd.concat([df_test,train_label], axis=1)
df_full = df_full.loc[:,~df_full.columns.duplicated()]
df_full.shape

df_full.SPENDINGRESPONSE.replace(["Spend to Improve Economy","Reduce National Debt and Deficit"], [1,0], inplace=True)

data_types = df_full.dtypes  
cat_cols = list(data_types[data_types=='object'].index)
con_cols = list(data_types[data_types=='int64'].index) + list(data_types[data_types=='float64'].index)

id_col = 'ID'
target_col = 'SPENDINGRESPONSE'
con_cols.remove('ID')
con_cols.remove('SPENDINGRESPONSE')

print("# of Categorical features:", len(cat_cols))
print("Categorical features:", cat_cols)
print("# of Numerical features:", len(con_cols))
print("Numerical features:", con_cols)
print("ID: %s, target: %s" %( id_col, target_col))

freq_encod_cat = []
for i in cat_cols:
    if df_full[i].value_counts().size >= 15:
        freq_encod_cat.append(i)
freq_encod_cat

## Feature combination

### Categorical features

cat_cols = [x for x in cat_cols if x not in freq_encod_cat]

import itertools
for comb in itertools.combinations(cat_cols, 2):
    feat = comb[0] + "_" + comb[1]
    df_full[feat] = df_full[comb[0]] + df_full[comb[1]]
    print('Combining Columns:', feat)

df_full.head()

data_types_2 = df_full.dtypes  
cat_cols_2 = list(data_types_2[data_types_2=='object'].index)

print(cat_cols_2, len(cat_cols_2))

cat_cols_2 = [x for x in cat_cols_2 if x not in freq_encod_cat]
cat_cols_2 = [x for x in cat_cols_2 if x not in cat_cols]

print(len(cat_cols_2))

LBL = preprocessing.LabelEncoder()
start = time.time()
LE_map = dict()
for cat_col in cat_cols_2:
    df_full[cat_col] = LBL.fit_transform(df_full[cat_col].astype(str))
    LE_map[cat_col]=dict(zip(LBL.classes_, LBL.transform(LBL.classes_)))
print ('Label enconding finished in %f seconds' % (time.time()-start))

OHE = preprocessing.OneHotEncoder(sparse=False)
start = time.time()
OHE.fit(df_full[cat_cols_2])
OHE_data = OHE.transform(df_full[cat_cols_2])
print ('One-hot-encoding finished in %f seconds' % (time.time()-start))

OHE_vars = [var + '_' + str(level).replace(' ','_')\
                for var in cat_cols_2 for level in LE_map[var]]

print (OHE_data.shape)

There is no obvious relationship among numerical features. And the feature importance from XGBoost shows that numerical features dominant in the model. The feature combination here will just limited to categorical features. 

df_full = pd.concat((df_full.ID, pd.DataFrame(OHE_data,columns=OHE_vars)), axis = 1)

df_full.to_csv('../data/input_cat_comb1.csv', index=False)
