==> Voter_Classifications/5_output.ipynb <==
# Classification outputs

import pandas as pd
import numpy as np
import xgboost as xgb
import seaborn as sns
from scipy.stats import skew, boxcox
from scipy import sparse
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report
from sklearn.model_selection import train_test_split
import time
import matplotlib.pyplot as plt
%matplotlib inline

Load probability data 

df_prob = pd.read_csv('../data/xgb_prob_opt1.csv')
df = pd.read_csv('../data/input_proc2.csv') # with original label
df = pd.concat([df_prob, df.SPENDINGRESPONSE], axis=1)

train_size = 20000
df.SPENDINGRESPONSE[train_size:].isnull().sum()

Comput accuracy to find the threshold of predicted probablity.  

Y = df.SPENDINGRESPONSE[:train_size]
X = df.Probability[:train_size]

X.describe()

Split dataset to train and test. Use train dataset to find the threshold of predicted probablity, and test the results on the test dataset.

x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, random_state=1234)
x_train_sort = sorted(x_train, reverse=True)

count=0
for i in y_train:
    if i==1:
        count +=1
threh = x_train_sort[count]
threh

y_pred = [1 if x > threh else 0 for x in x_test]

Performance measures

print(classification_report(y_pred, y_test))

print(accuracy_score(y_pred, y_test))

print(confusion_matrix(y_pred, y_test))

Generate final submission.

y_class_full = [1 if x > threh else 0 for x in df.Probability]
len(y_class_full)

y_class_full_word = ["Spend to Improve Economy" if x==1 else "Reduce National Debt and Deficit" for x in y_class_full]

fnl_sub_df = pd.DataFrame({'ID': df.ID, 'Probability': df.Probability, 'Classification_1_0': y_class_full, 'Classification': y_class_full_word})
fnl_sub_df = fnl_sub_df[['ID','Probability','Classification_1_0','Classification']]
fnl_sub_df.to_csv("../data/final_submission.csv", index=False)
