==> Voter_Classifications/6_output_visual.ipynb <==
# Output Visualization

import pandas as pd
import numpy as np
import xgboost as xgb
import seaborn as sns
from scipy.stats import skew, boxcox
from scipy import sparse
from sklearn.metrics import confusion_matrix, accuracy_score, classification_report
from sklearn.model_selection import train_test_split
import time
import matplotlib.pyplot as plt
%matplotlib inline

df_fnl = pd.read_csv('../data/final_submission.csv')
df_ipt = pd.read_csv('../data/input_proc2.csv') # with original label
test = pd.read_csv('../ProjectFiles/File3.csv')

df = pd.concat([df_ipt, test.State, df_fnl.Classification_1_0], axis=1)

col = ['State','f9','f84','f89','f29','f83', 'Classification_1_0']
df = df[col]

df.head()

df_mean = df.groupby(['State'], as_index=False).mean()
df_mean.head(10)

df_mean.to_csv("../data/output_visual.csv", index=False)
